requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
supabase==2.4.0
python-dotenv==1.0.1
whitenoise==6.6.0